+++ src/dexterity/localrolesfield/testing.py
@@ -1,22 +1,30 @@
 # encoding: utf-8
-from zope.interface import alsoProvides
+import dexterity.localrolesfield as package
 from dexterity.localrolesfield.field import LocalRolesField, LocalRoleField
+
+import pkg_resources
 from plone import api
-from plone.autoform.interfaces import IFormFieldProvider
 from plone.app.robotframework.testing import AUTOLOGIN_LIBRARY_FIXTURE
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PloneWithPackageLayer
 from plone.app.testing import ploneSite
+from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.content import Container
 from plone.supermodel import model
 from plone.testing import z2
 from zope import interface
+from zope.interface import alsoProvides
 from zope.schema import Choice
 from zope.schema.fieldproperty import FieldProperty
 from zope.schema.vocabulary import SimpleVocabulary
 
-import dexterity.localrolesfield as package
+try:
+    pkg_resources.get_distribution('plone.app.contenttypes')
+except pkg_resources.DistributionNotFound:
+    HAS_PA_CONTENTTYPES = False
+else:
+    HAS_PA_CONTENTTYPES = True
 
 
 class ITestingType(model.Schema):
@@ -78,6 +86,13 @@ class LocalRolesFieldLayer(PloneWithPackageLayer):
             if not api.user.get(username='basic-user'):
                 api.user.create(username='basic-user', email='test@test.com')
 
+    def setUpPloneSite(self, portal):
+        """Set up Plone."""
+        super(LocalRolesFieldLayer, self).setUpPloneSite(portal)
+
+        # Plone 5 support
+        if HAS_PA_CONTENTTYPES:
+            self.applyProfile(portal, 'plone.app.contenttypes:default')
 
 LOCALROLESFIELD_FIXTURE = LocalRolesFieldLayer(
     zcml_filename='testing.zcml',